Implemented all of Project and its tests

+++ mv_loader/migrations/20240805032951_Projects.sql
@@ -4,8 +4,8 @@ CREATE TABLE IF NOT EXISTS Projects (
   ProjectId       GUID NOT NULL UNIQUE,
   ProjectName     VARCHAR(255) NOT NULL,
   ProjectDate     DATE NOT NULL,
-  PayRate         DECIMAL(10, 2) NOT NULL,
+  PayRate         REAL NOT NULL,
   ProjectDuration INTEGER NOT NULL,    -- The duration in milliseconds
-  TotalPay        DECIMAL(10, 2) NOT NULL,
+  TotalPay        REAL NOT NULL,
   CONSTRAINT pk_Projects PRIMARY KEY(ProjectID)
 );
